+++ packages/react-native-sdk/src/components/Participant/ParticipantView/VideoRenderer.tsx
@@ -73,7 +73,8 @@ export const VideoRenderer = ({
     deregister: deregisterIosScreenshot,
   } = useScreenshotIosContext();
 
-  const videoDimensions = useTrackDimensions(participant, trackType);
+  const videoDimensions = useTrackDimensions(participant);
+
   const {
     isLocalParticipant,
     sessionId,